+++ images/common/harness/run_harness.rb
@@ -17,13 +17,8 @@ class RunHarness < BaseHarness
     raise 'abstract RunHarness#program'
   end
 
-  def stdin
-    input_json['stdin']
-  end
-
   def validate_input_json
     abort 'No program given.' if !program
-    abort 'No stdin given.' if !stdin
   end
 
 end