+++ sandboxes/common/test/lib/input-json-for-compile.bash
@@ -3,5 +3,5 @@
 function input_json {
   set -e
   var_is_set source_program
-  jq -n --arg sp "$source_program" '{sourceProgram: $sp}'
+  printf '%s' "$source_program" | jq -R --slurp '{sourceProgram: .}'
 }